requirements.txt
alembic==1.4.2
ase==3.19.1
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
configparser==5.0.0
coverage==5.0
cycler==0.10.0
databricks-cli==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docker==4.2.0
entrypoints==0.3
Flask==1.1.2
future==0.18.2
gitdb==4.0.5
GitPython==3.1.2
googledrivedownloader==0.4
gorilla==0.3.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
imageio==2.8.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1593446408836/work
ipykernel==5.3.0
ipython @ file:///tmp/build/80754af9/ipython_1593447368578/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi @ file:///tmp/build/80754af9/jedi_1592841891421/work
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
llvmlite==0.32.0
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mlflow==1.8.0
mock==4.0.2
more-itertools==8.4.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numba==0.49.0
numpy==1.16.4
olefile==0.46
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
plyfile==0.7.2
prometheus-client==0.8.0
prometheus-flask-exporter==0.13.0
prompt-toolkit==3.0.5
protobuf==3.12.1
ptyprocess==0.6.0
py @ file:///tmp/build/80754af9/py_1593446248552/work
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
pytest-cov==2.8.1
pytest-runner==5.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole @ file:///tmp/build/80754af9/qtconsole_1592848611704/work
QtPy==1.9.0
querystring-parser==1.2.4
rdflib==5.0.0
requests==2.23.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
simplejson==3.17.0
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.13
sqlparse==0.3.1
tabulate==0.8.7
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
torch-cluster==1.5.4
torch-geometric==1.4.3
torch-scatter==2.0.4
torch-sparse==0.6.1
torch-spline-conv==1.2.0
torchsummary==1.5.1
torchvision==0.2.1
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
trimesh==3.6.43
urllib3==1.25.9
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.1.0
